Also update KDF cache in biometric storage if password changes
This fixes various bugs including several relating to the uninstallation
 and reinstallation of the app on iOS, particularly affecting the
autofill feature because that relies on the KDF result since it does
not have the memory capacity to run an argon2 derivation on demand.

Will also fix up slower unlocking on Android if user had changed
their password since we introduced the KDF cache.

Also resets the countdown to the next biometric sign in being required
when a new password is used to unlock the vault.